Move components to packages, display suggestions only after 2 chars
