+++ client/src/resources/organizations/OrganizationShow.js
@@ -1,8 +1,7 @@
 import React from 'react';
 import { TextField, UrlField } from 'react-admin';
-import { Column, ColumnShowLayout, Hero, UserIcon, GridList, Show } from '../../archipelago-layout';
-import MarkDownField from '../../components/MarkdownField';
-import UriArrayField from '../../components/UriArrayField';
+import { Column, ColumnShowLayout, Hero, UserIcon, GridList, Show, MarkdownField } from '../../archipelago-layout';
+import { UriArrayField } from '../../semantic-data-provider';
 
 const OrganizationTitle = ({ record }) => {
   return <span>{record ? record['pair:label'] : ''}</span>;
@@ -16,7 +15,7 @@ const OrganizationShow = props => (
           <TextField label="Courte description" source="pair:comment" />
           <UrlField label="Site web" source="pair:homePage" />
         </Hero>
-        <MarkDownField source="pair:description" addLabel />
+        <MarkdownField source="pair:description" addLabel />
       </Column>
       <Column xs={12} sm={3} showLabel>
         <UriArrayField label="Membres" reference="User" source="pair:hasMember" referenceBasePath="/User">
