+++ client/src/semantic-data-provider/index.js
@@ -1,2 +1,4 @@
 export { default as dataProvider } from './dataProvider';
 export { default as httpClient } from './httpClient';
+export { default as UriArrayField } from './fields/UriArrayField';
+export { default as UriArrayInput } from './inputs/UriArrayInput';
